+++ app/src/main/java/ru/netology/nmedia/dto/Post.kt
@@ -5,7 +5,7 @@ data class Post(
     val author: String,
     val content: String,
     val published: String,
-    var likes: Int = 999,
+    var likes: Int = 9,
     var likedByMe: Boolean = false,
     var repost: Int = 0,
     var views: Int = 100
