+++ app/src/main/res/drawable/ic_netology_48dp.xml
@@ -1,0 +1,21 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="48dp"
+    android:height="48dp"
+    android:viewportWidth="48"
+    android:viewportHeight="48">
+  <path
+      android:pathData="M36.5739,48C30.2635,48 25.1478,42.8844 25.1478,36.5739C25.1478,30.2635 30.2635,25.1478 36.5739,25.1478C42.8844,25.1478 48,30.2635 48,36.5739C48,42.8844 42.8844,48 36.5739,48Z"
+      android:fillColor="#0066FF"/>
+  <path
+      android:pathData="M11.4261,48C17.7365,48 22.8522,42.8844 22.8522,36.5739C22.8522,30.2635 17.7365,25.1478 11.4261,25.1478C5.1156,25.1478 0,30.2635 0,36.5739C0,42.8844 5.1156,48 11.4261,48Z"
+      android:fillColor="#5D00F5"/>
+  <path
+      android:pathData="M11.4261,22.8522C17.7365,22.8522 22.8522,17.7365 22.8522,11.4261C22.8522,5.1156 17.7365,0 11.4261,0C5.1156,0 0,5.1156 0,11.4261C0,17.7365 5.1156,22.8522 11.4261,22.8522Z"
+      android:fillColor="#EB236B"/>
+  <path
+      android:pathData="M36.5739,22.8522C42.8844,22.8522 48,17.7365 48,11.4261C48,5.1156 42.8844,0 36.5739,0C30.2635,0 25.1478,5.1156 25.1478,11.4261C25.1478,17.7365 30.2635,22.8522 36.5739,22.8522Z"
+      android:fillColor="#4BD0A0"/>
+  <path
+      android:pathData="M23.7391,35.3739C30.0496,35.3739 35.1652,30.2349 35.1652,23.8956C35.1652,17.5564 30.0496,12.4174 23.7391,12.4174C17.4287,12.4174 12.313,17.5564 12.313,23.8956C12.313,30.2349 17.4287,35.3739 23.7391,35.3739Z"
+      android:fillColor="#ffffff"/>
+</vector>
